--- generators/app/templates/app/components/CarsList/CarsList.scss
@@ -1,3 +1,0 @@
-.CarsList {
-
-}
