+++ generators/app/templates/app/containers/Account/Account.js
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import Actions from '../../actions'
+import * as Actions from '../../actions'
 import { Link } from 'react-router'
 import './Account.scss'
