+++ generators/app/templates/app/containers/App/App.js
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import Actions from '../../actions'
+import * as Actions from '../../actions'
 import Navbar from '../../components/Navbar/Navbar'
 import './App.scss'
 
@@ -15,7 +15,7 @@ class Main extends Component {
       username: PropTypes.string
     }),
     logout: PropTypes.func
-  };
+  }
 
   render () {
     return (
@@ -29,6 +29,7 @@ class Main extends Component {
     )
   }
 }
+
 // Place state of redux store into props of component
 function mapStateToProps (state) {
   return {
@@ -36,6 +37,7 @@ function mapStateToProps (state) {
     router: state.router
   }
 }
+
 // Place action methods into props
 function mapDispatchToProps (dispatch) {
   return bindActionCreators(Actions, dispatch)
