+++ generators/app/templates/fireuser/configureStore.dev.js
@@ -1,0 +1,24 @@
+import { createStore, applyMiddleware, compose } from 'redux'
+import rootReducer from '../reducers'
+import thunk from 'redux-thunk'
+import { Middleware } from 'redux-devshare'
+import { syncHistory } from 'react-router-redux'
+
+export default function configureStore (initialState, history) {
+  const reduxRouterMiddleware = syncHistory(history)
+  const createStoreWithMiddleware = compose(
+    applyMiddleware(thunk, Middleware, reduxRouterMiddleware),
+    typeof window === 'object' && typeof window.devToolsExtension !== 'undefined' ? window.devToolsExtension() : f => f
+  )(createStore)
+  const store = createStoreWithMiddleware(rootReducer, initialState)
+
+  if (module.hot) {
+    // Enable Webpack hot module replacement for reducers
+    module.hot.accept('../reducers', () => {
+      const nextRootReducer = require('../reducers')
+      store.replaceReducer(nextRootReducer)
+    })
+  }
+
+  return store
+}
