+++ generators/app/templates/fireuser/index-server.js
