+++ generators/app/templates/fireuser/reducers.js
@@ -1,13 +1,15 @@
 import { combineReducers } from 'redux'
-import { routerStateReducer } from 'redux-router'
-import cars from './cars'
+import { routeReducer } from 'react-router-redux'
 import { Reducers } from 'redux-fireuser'
-const { account, entities } = Reducers
+import tabs from './tabs'
+const { account, projects, entities } = Reducers
+
 const rootReducer = combineReducers({
-  cars,
   account,
   entities,
-  router: routerStateReducer
+  tabs,
+  projects,
+  router: routeReducer
 })
 
 export default rootReducer
